requirements.txt
fastmcp==2.3.4
google-cloud-bigquery==3.27.0
google-auth==2.38.0
starlette==0.46.1
uvicorn==0.34.0